moved google analytics functionality to _app
