+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
   "extends": [
    "next/core-web-vitals",
-   "eslint:recommended",
+   "next",
    "prettier"
   ],
   "plugins":[],
