+++ components/Head/Head.tsx
@@ -1,17 +1,20 @@
 import NextHead from "next/head";
-import { useRouter } from "next/router"
+import { useRouter } from "next/router";
 import React from "react";
 
 const Head = () => {
   const router = useRouter();
-  const url = router.asPath
+  const url = router.asPath;
 
   return (
     <NextHead>
       <title>downtime.dev</title>
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
       <link rel="icon" href="/assets/favicon.ico" />
-      <meta name="description" content="Hard-hitting tech news while your code compiles." />
+      <meta
+        name="description"
+        content="Hard-hitting tech news while your code compiles."
+      />
       <style>
         @import
         url(&apos;https://fonts.googleapis.com/css2?family=Ubuntu+Mono:ital,wght@0,400;0,700;1,400;1,700&display=swap&apos;);{" "}
@@ -19,13 +22,25 @@ const Head = () => {
       <meta property="og:type" content="website" />
       <meta property="og:url" content={url} />
       <meta property="og:title" content="downtime.dev" />
-      <meta property="og:description" content= "Hard-hitting news for when your code is compiling." />
-      <meta property="og:image" content="https://www.downtime.dev/assets/images/bored@2x.jpg" />
+      <meta
+        property="og:description"
+        content="Hard-hitting news for when your code is compiling."
+      />
+      <meta
+        property="og:image"
+        content="https://www.downtime.dev/assets/images/bored@2x.jpg"
+      />
       <meta name="twitter:card" content="summary" />
       <meta name="twitter:title" content="downtime.dev" />
-      <meta property="twitter:description" content= "Hard-hitting news for when your code is compiling." />
+      <meta
+        property="twitter:description"
+        content="Hard-hitting news for when your code is compiling."
+      />
       <meta property="twitter:url" content={url} />
-      <meta property="twitter:image" content="https://www.downtime.dev/assets/images/bored@2x.jpg" />
+      <meta
+        property="twitter:image"
+        content="https://www.downtime.dev/assets/images/bored@2x.jpg"
+      />
     </NextHead>
   );
 };
