+++ components/Head/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Head"
+export { default } from "./Head";
