+++ components/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import NavBar from "components/NavBar/NavBar"
-import GDPR from "components/GDPR"
+import NavBar from "components/NavBar/NavBar";
+import GDPR from "components/GDPR";
 
 interface LayoutProps {
   children: React.ReactNode;
@@ -9,11 +9,11 @@ interface LayoutProps {
 const Layout = ({ children }: LayoutProps) => {
   return (
     <div>
-      <NavBar/>
+      <NavBar />
       <GDPR />
       {children}
     </div>
-    )
+  );
 };
 
 export default Layout;
