+++ components/Layout/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Layout"
+export { default } from "./Layout";
