+++ components/NavBar/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./NavBar"
+export { default } from "./NavBar";
