+++ pages/_app.tsx
@@ -4,14 +4,16 @@ import type { AppProps } from "next/app";
 import Layout from "components/Layout";
 import Head from "components/Head";
 import "../styles/globals.css";
-import { pageview } from "utilities/google";
+import { pageview } from "utilities/google/gtag";
+import Script from "next/script";
 
 function MyApp({ Component, pageProps }: AppProps) {
   const router = useRouter();
 
   useEffect(() => {
     const handleRouteChange = (url: any) => {
-      pageview(url); 
+      //utility for logging page views in Google Analytics
+      pageview(url);
     };
     //When the component is mounted, subscribe to router changes
     //and log those page views
@@ -26,6 +28,26 @@ function MyApp({ Component, pageProps }: AppProps) {
 
   return (
     <>
+      {/* Google Analytics Start */}
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GTAG_ID}`}
+      />
+      <Script
+        id="gtag-init"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', '${process.env.NEXT_PUBLIC_GTAG_ID}', {
+              page_path: window.location.pathname,
+            });
+          `,
+        }}
+      />
+      {/* Google Analytics End */}
       <Head />
       <Layout>
         <Component {...pageProps} />
