+++ pages/_document.tsx
@@ -15,24 +15,7 @@ class MyDocument extends Document {
   render() {
     return (
       <Html>
-        <Head>
-          <script
-            async
-            src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GTAG_ID}`}
-          />
-          <script
-            dangerouslySetInnerHTML={{
-              __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', '${process.env.NEXT_PUBLIC_GTAG_ID}', {
-              page_path: window.location.pathname,
-            });
-          `,
-            }}
-          />
-        </Head>
+        <Head></Head>
         <body>
           <Main />
           <NextScript />
