+++ pages/terms.tsx
