+++ utilities/google/gtag.ts
@@ -1,0 +1,8 @@
+
+// https://developers.google.com/analytics/devguides/collection/gtagjs/pages
+export const pageview = (url: URL) => {
+  //@types gtag.js modified so that targetId = string | undefined (added undefined) - Cole
+  window.gtag('config', process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS, {
+    page_path: url,
+  })
+}
