--- utilities/google/index.js
@@ -1,5 +1,0 @@
-export const pageview = (url) => {
-  window.gtag('config', process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS, {
-    page_path: url,
-  })
-}
